+++ 微球/WBQuestionsViewController.m
@@ -288,7 +288,7 @@
     self.selectedRow = (int)cell.tag;
     WBQuestionsListModel *data = self.questionsList[self.selectedRow];
     WBHomepageViewController *homepage = [[WBHomepageViewController alloc] init];
-    homepage.userId = [NSString stringWithFormat:@"%ld",data.hga.tblUser.userId];
+    homepage.userId = [NSString stringWithFormat:@"%ld",(long)data.hga.tblUser.userId];
     homepage.hidesBottomBarWhenPushed = YES;
     if (self.fromFindView) {
         [self.parentViewController.navigationController pushViewController:homepage animated:YES];
