+++ 微球/WBTopicDetailViewController.m
@@ -178,7 +178,7 @@
         
         [parameters setObject:[WBUserDefaults userId] forKey:@"userId"];
         [parameters setObject:@"2" forKey:@"newsType"];
-        [parameters setObject:[NSString stringWithFormat:@"%ld",_topicID] forKey:@"topicId"];
+        [parameters setObject:[NSString stringWithFormat:@"%ld",(long)_topicID] forKey:@"topicId"];
         
         [MyDownLoadManager postUserInfoUrl:@"http://121.40.132.44:92/tq/setComment" withParameters:parameters fieldData:^(id<AFMultipartFormData> formData) {
             [formData appendPartWithFileData:fileData name:@"1234" fileName:@"video1.mov" mimeType:@"video/quicktime"];
@@ -319,7 +319,7 @@
     
     WBTopicCommentTableViewController *commentView = [[WBTopicCommentTableViewController alloc]init];
     commentView.commentId =((TopicDetailModel *)_dataArray[indexPath.row]).commentId;
-    commentView.userId =[NSString stringWithFormat:@"%ld", ((TopicDetailModel *)_dataArray[indexPath.row]).userId];
+    commentView.userId =[NSString stringWithFormat:@"%ld", (long)((TopicDetailModel *)_dataArray[indexPath.row]).userId];
     
     [self.navigationController pushViewController:commentView animated:YES];
